requirements.txt
fastapi
uvicorn
sqlalchemy
langgraph
langchain
chromadb
ollama
python-multipart
bcrypt==4.0.1
sqlite-utils
pdfplumber
pydantic
python-jose[cryptography]
aiofiles
pillow
httpx
passlib
python-dotenv
langchain-community
langchain-ollama